mastodon: Cache root toots in redis for 4 weeks
Searching the initial root toot is the most expensive task of the
script. Only caching it as part of the cached comments means we
only cache it for 10 minutes, which is not a lot.

Cache it for 4 weeks instead if we found a root toot. If we did
not find a root toot, cache that fact for 10 minutes. Probably
want to improve on that, so we can cache the 404 for 4 weeks as
well for posts imported from old blogs without comments on mastodon.