lazy load avatars when interactionobserver/isIntersecting is available
This improves page load performance, as we are not loading
avatars when not showing comments. The avatars do have a fixed
size, so loading them lazily also does not affect scrolling or
layout.

Another alternative is to lazy load the entire comments part,
but this does affect scrolling, as the document becomes longer,
and hence might reduce the experience.